web/requirements.txt

Django==5.1.6
psycopg2-binary
python-dotenv==1.0.1
gunicorn==23.0.0
django-storages==1.14.5
django_csp==3.8
django-sri==0.8.0
cryptography==43.0.3
django-environ
requests
djangorestframework
djangorestframework-simplejwt
huggingface_hub
google-genai